fix: CORS configuration, security filters, and workflow UI improvements
- Fixed CORS preflight handling in RateLimitingFilter (removed CorsConfig.java)
- Enhanced SecurityConfig with proper CORS headers and CSRF exemptions
- Added X-Forwarded-For support for rate limiting
- Improved JWT authentication filter
- Updated workflow page UI enhancements
- Fixed HealthConfig for notification service
- Updated application-local.yml configuration
- Various service improvements

+++ jira-plan-service/src/main/java/com/jira/plan/service/SprintSnapshotService.java
@@ -260,7 +260,7 @@ public class SprintSnapshotService {
     private BigDecimal getIssuePoints(SprintIssue issue) {
         // Try to get points from plan item first
         if (issue.getPlanItem() != null && issue.getPlanItem().getStoryPoints() != null) {
-            return issue.getPlanItem().getStoryPoints();
+            return BigDecimal.valueOf(issue.getPlanItem().getStoryPoints());
         }
 
         // Points from story points field
